refactor: move wheel_builder error type as it was growing large
